Modified code to confirm with proper(?) listener syntax.
I have not yet worked with the listner interface so this change might
be incorrect. In getting the example to run as I expected I did some minor
modification to the code and have included some documentation on how
I run it. Please correct me if I am heading down the wrong path with these
changes.